feat: implement notifications context and provider for better state management
