+++ src/hooks/useNotifications.tsx
@@ -1,0 +1,98 @@
+import { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react";
+import { invoke } from "@tauri-apps/api/core";
+import type { NotificationResult } from "../types";
+import { useI18n } from "../i18n";
+
+interface NotificationsContextValue {
+  result: NotificationResult | null;
+  loading: boolean;
+  error: string | null;
+  fetchNotifications: () => Promise<void>;
+  markRead: (id: string) => Promise<void>;
+  markAllRead: () => Promise<void>;
+}
+
+const NotificationsContext = createContext<NotificationsContextValue | null>(null);
+
+export function NotificationsProvider({ children }: { children: React.ReactNode }) {
+  const { t } = useI18n();
+  const [result, setResult] = useState<NotificationResult | null>(null);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const fetchNotifications = useCallback(async () => {
+    setLoading(true);
+    try {
+      const data = await invoke<NotificationResult>("get_notifications");
+      setResult(data);
+      setError(null);
+    } catch (err) {
+      const message =
+        typeof err === "string"
+          ? err
+          : err instanceof Error
+            ? err.message
+            : t("notification.loadingFailed");
+      setError(message);
+      console.error("Failed to load notifications:", err);
+    } finally {
+      setLoading(false);
+    }
+  }, [t]);
+
+  useEffect(() => {
+    fetchNotifications();
+  }, [fetchNotifications]);
+
+  const markRead = useCallback(async (id: string) => {
+    try {
+      await invoke("mark_notification_read", { id });
+      setResult((prev) => {
+        if (!prev) return prev;
+        const notifications = prev.notifications.map((n) =>
+          n.id === id ? { ...n, isRead: true } : n,
+        );
+        const unreadCount = notifications.filter((n) => !n.isRead).length;
+        const hasUnreadPopup = notifications.some((n) => !n.isRead && n.popup);
+        return { notifications, unreadCount, hasUnreadPopup };
+      });
+    } catch (err) {
+      console.error("Failed to mark notification as read:", err);
+    }
+  }, []);
+
+  const markAllRead = useCallback(async () => {
+    try {
+      await invoke("mark_all_notifications_read");
+      setResult((prev) => {
+        if (!prev) return prev;
+        const notifications = prev.notifications.map((n) => ({ ...n, isRead: true }));
+        return { notifications, unreadCount: 0, hasUnreadPopup: false };
+      });
+    } catch (err) {
+      console.error("Failed to mark all notifications as read:", err);
+    }
+  }, []);
+
+  const value = useMemo(
+    () => ({
+      result,
+      loading,
+      error,
+      fetchNotifications,
+      markRead,
+      markAllRead,
+    }),
+    [error, fetchNotifications, loading, markAllRead, markRead, result],
+  );
+
+  return <NotificationsContext.Provider value={value}>{children}</NotificationsContext.Provider>;
+}
+
+export function useNotifications() {
+  const context = useContext(NotificationsContext);
+  if (!context) {
+    throw new Error("useNotifications must be used within NotificationsProvider");
+  }
+  return context;
+}
