+++ src/main.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import { ToastProvider } from "./components/Toast";
+import { NotificationsProvider } from "./hooks/useNotifications";
 import { I18nProvider } from "./i18n";
 
 class ErrorBoundary extends React.Component<
@@ -62,7 +63,9 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
     <ErrorBoundary>
       <I18nProvider>
         <ToastProvider>
-          <App />
+          <NotificationsProvider>
+            <App />
+          </NotificationsProvider>
         </ToastProvider>
       </I18nProvider>
     </ErrorBoundary>
